+++ include/provenanceutils.h
@@ -36,6 +36,7 @@ static const char RL_STR_WRITE []                 = "write";
 static const char RL_STR_CREATE []                = "create";
 static const char RL_STR_CHANGE []                = "change";
 static const char RL_STR_MMAP_WRITE []            = "mmap_write";
+static const char RL_STR_SH_WRITE []              = "sh_write";
 static const char RL_STR_BIND []                  = "bind";
 static const char RL_STR_CONNECT []               = "connect";
 static const char RL_STR_LISTEN []                = "listen";
@@ -59,6 +60,7 @@ static const char RL_STR_GETXATTR []              = "getxattr";
 static const char RL_STR_LSTXATTR []              = "listxattr";
 static const char RL_STR_READLINK []              = "readlink";
 static const char RL_STR_MMAP_READ []             = "mmap_read";
+static const char RL_STR_SH_READ []               = "sh_read";
 static const char RL_STR_MMAP_EXEC []             = "mmap_exec";
 static const char RL_STR_SND []                   = "send";
 static const char RL_STR_SND_PACKET []            = "send_packet";
@@ -138,6 +140,10 @@ static inline const char* relation_str(uint64_t type){
       return RL_STR_PERM_READ;
     case RL_PERM_WRITE:
       return RL_STR_PERM_WRITE;
+    case RL_SH_READ:
+      return RL_STR_SH_READ;
+    case RL_SH_WRITE:
+      return RL_STR_SH_WRITE;
     case RL_PERM_EXEC:
       return RL_STR_PERM_EXEC;
     default:
@@ -184,6 +190,8 @@ static inline const uint64_t relation_id(char* str){
   MATCH_AND_RETURN(str, RL_STR_PERM_READ, RL_PERM_READ);
   MATCH_AND_RETURN(str, RL_STR_PERM_WRITE, RL_PERM_WRITE);
   MATCH_AND_RETURN(str, RL_STR_PERM_EXEC, RL_PERM_EXEC);
+  MATCH_AND_RETURN(str, RL_STR_SH_READ, RL_SH_READ);
+  MATCH_AND_RETURN(str, RL_STR_SH_WRITE, RL_SH_WRITE);
   return 0;
 }